+++ tsconfig.json
@@ -3,6 +3,7 @@
     "jsx": "react",
     "jsxFactory": "jsx",
 
+    "target": "ES2022",
     "outDir": "prebuild/js",
     "module": "ESNext",
     "moduleResolution": "Node"
